+++ gaelach/__init__.py
@@ -48,10 +48,10 @@ from gaelach.utils.if_else import if_else
 from gaelach.utils.case_when import case_when
 from gaelach.utils.row_contains import row_contains
 from gaelach.utils.lambdas import to_lower, to_upper, to_strip, to_title, to_str, to_int, \
-   to_bool, to_date, to_na, to_zero, to_round
+   to_float, to_date, to_na, to_zero, to_round, to_cat
     
 # set version
-__version__ = "0.2.0"
+__version__ = "0.2.1"
 
 # add to primary import 
 __all__ = ['_', 'Symbolic', 'select', 'mutate', 'filter', 'across', 'where', 'is_boolean', 
@@ -61,6 +61,6 @@ __all__ = ['_', 'Symbolic', 'select', 'mutate', 'filter', 'across', 'where', 'is
            'pivot_longer', 'pivot_wider', 'unite', 'separate', 'bind_rows', 'bind_cols', 
            'arrange', 'distinct', 'head', 'tail', 'drop_na', 'slice', 'sample', 'rename', 
            'round', 'relocate', 'drop_na', 'if_else', 'case_when', 'row_contains', 'to_lower', 
-           'to_upper', 'to_strip', 'to_title', 'to_str', 'to_int', 'to_bool', 'to_date', 'to_na', 
-           'to_zero', 'to_round'
+           'to_upper', 'to_strip', 'to_title', 'to_str', 'to_int', 'to_float', 'to_date', 
+           'to_na', 'to_zero', 'to_round', 'to_cat'
         ]